Update view permissions

+++ services/views.py
@@ -6,6 +6,7 @@ from .models import Tender, Input, Investment
 from rest_framework.permissions import IsAuthenticated
 from django.contrib.auth import get_user_model
 from drf_yasg.utils import swagger_auto_schema
+from .permisisions import IsOwnerOnly
 
 User=get_user_model()
 
@@ -19,92 +20,120 @@ class HelloServicesView(generics.GenericAPIView):
 class TenderCreateListView(generics.GenericAPIView):
     serializer_class=serializers.TenderViewSerializer
     queryset=Tender.objects.all()
-    permission_classes=[IsAuthenticated]
+    permission_classes=[IsAuthenticated, IsOwnerOnly]
 
     @swagger_auto_schema(operation_summary="list all tenders")
     def get(self, request):
+        if request.user.is_tender_holder == True:
 
-        tenders = Tender.objects.all()
+            tenders = Tender.objects.all()
 
-        serializer=self.serializer_class(instance=tenders, many=True)
+            serializer=self.serializer_class(instance=tenders, many=True)
+        else:
+            return Response({"Error":"User must be an tender user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
  
     @swagger_auto_schema(operation_summary="create a new tender")   
     def post (self, request):
-        data=request.data
+        if request.user.is_tender_holder == True:
 
-        serializer=self.serializer_class(data=data)
+            data=request.data
 
-        user=request.user
+            serializer=self.serializer_class(data=data)
 
-        if serializer.is_valid():
-            serializer.save(tender_holder=user)
+            user=request.user
 
-            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
-        
+            if serializer.is_valid():
+                serializer.save(tender_holder=user)
+
+                return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response({"Error":"User must be an tender user"})
+            
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 class TenderDetailView(generics.GenericAPIView):
     serializer_class=serializers.TenderDetailViewSerializer
-    #permission_classes=[IsAuthenticated]
+    permission_classes=[IsAuthenticated,]
 
     @swagger_auto_schema(operation_summary="Retrieve a tender by id")
     def get(self, request, tender_id):
-        
-        tender=get_object_or_404(Tender, pk=tender_id)
+        if request.user.is_tender_holder == True:
 
-        serializer=self.serializer_class(instance=tender)
+            tender=get_object_or_404(Tender, pk=tender_id)
+
+            serializer=self.serializer_class(instance=tender)
+        else:
+            return Response({"Error":"User must be an tender user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     @swagger_auto_schema(operation_summary="Update a tender by id")
     def put(self, request, tender_id):
-        data=request.data 
+        if request.user.is_tender_holder == True:
+            data=request.data 
 
-        tender=get_object_or_404(Tender, pk=tender_id)
+            tender=get_object_or_404(Tender, pk=tender_id)
 
-        serializer=self.serializer_class(data=data, instance=tender)
+            serializer=self.serializer_class(data=data, instance=tender)
 
-        if serializer.is_valid():
-            serializer.save()
+            if serializer.is_valid():
+                serializer.save()
 
-            return Response(data=serializer.data, status=status.HTTP_200_OK)
-        
+                return Response(data=serializer.data, status=status.HTTP_200_OK)
+        else:
+            return Response({"Error":"User must be an tender user"})
+            
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         
     @swagger_auto_schema(operation_summary="Delete a tender")
     def delete(self, request, tender_id):
-        tender=get_object_or_404(Tender, pk=tender_id)
+        if request.user.is_tender_holder == True:
+            tender=get_object_or_404(Tender, pk=tender_id)
 
-        tender.delete()
+            tender.delete()
+        else:
+            return Response({"Error":"User must be an tender user"})
 
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 class UserTendersView(generics.GenericAPIView):
     serializer_class=serializers.TenderDetailViewSerializer
+    permission_classes=[IsAuthenticated,]
+
 
     @swagger_auto_schema(operation_summary="Get all tenders for a user")
-    def get(self, request, user_id): 
-        user=User.objects.get(pk=user_id)
+    def get(self, request, user_id):
+        if request.user.is_tender_holder == True: 
+            user=User.objects.get(pk=user_id)
 
-        tenders=Tender.objects.all().filter(tender_holder=user)
+            tenders=Tender.objects.all().filter(tender_holder=user)
 
-        serializer=self.serializer_class(instance=tenders, many=True)
+            serializer=self.serializer_class(instance=tenders, many=True)
+        else:
+            return Response({"Error":"User must be a tender user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
 
 class UserTenderDetail(generics.GenericAPIView):
     serializer_class=serializers.TenderDetailViewSerializer
+    permission_classes=[IsAuthenticated,]
+
 
     @swagger_auto_schema(operation_summary="Get a user's specific tender")
     def get(self, request, user_id, tender_id):
-        user=User.objects.get(pk=user_id)
+        if request.user.is_tender_holder == True:
+            
+            user=User.objects.get(pk=user_id)
 
-        tenders=Tender.objects.all().filter(tender_holder=user).get(pk=tender_id)
+            tenders=Tender.objects.all().filter(tender_holder=user).get(pk=tender_id)
 
-        serializer=self.serializer_class(instance=tenders)
+            serializer=self.serializer_class(instance=tenders)
+        
+        else:
+            return Response({"Error":"User must be an tender user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
@@ -115,93 +144,121 @@ class InputCreateListView(generics.GenericAPIView):
     queryset=Input.objects.all()
     permission_classes=[IsAuthenticated]
 
+    
     @swagger_auto_schema(operation_summary="List all inputs")
     def get(self, request):
+        if request.user.is_input_holder == True:
+            inputs = Input.objects.all()
 
-        inputs = Input.objects.all()
-
-        serializer=self.serializer_class(instance=inputs, many=True)
-
-        return Response(data=serializer.data, status=status.HTTP_200_OK)
+            serializer=self.serializer_class(instance=inputs, many=True)
+            response=serializer.data
+            status_code=status.HTTP_200_OK
+        else:
+            return Response({"Error":"User must be an input user"})
+                #return Response(data=serializer.data, status=status.HTTP_200_OK)
+       
+        return Response(response, status_code)
 
     @swagger_auto_schema(operation_summary="Create a new input")
     def post (self, request):
-        data=request.data
+        if request.user.is_input_holder == True:
+            data=request.data
 
-        serializer=self.serializer_class(data=data)
+            serializer=self.serializer_class(data=data)     
 
-        user=request.user
+            user=request.user
 
-        if serializer.is_valid():
-            serializer.save(input_holder=user)
+            if serializer.is_valid():
+                serializer.save(input_holder=user)
 
-            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+                return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response({"Error":"User must be an input user"})
         
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class InputDetailView(generics.GenericAPIView):
     serializer_class=serializers.InputDetailViewSerializer
-    #permission_classes=[IsAuthenticated]
+    permission_classes=[IsAuthenticated]
 
     @swagger_auto_schema(operation_summary="Retrieve an input by id")
     def get(self, request, input_id):
+        if request.user.is_input_holder == True:
         
-        input=get_object_or_404(Input, pk=input_id)
+            input=get_object_or_404(Input, pk=input_id)
 
-        serializer=self.serializer_class(instance=input)
-
+            serializer=self.serializer_class(instance=input)
+        else:
+            return Response({"Error":"User must be an input user"})
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     @swagger_auto_schema(operation_summary="Update an input by id")
     def put(self, request, input_id):
-        data=request.data 
+        if request.user.is_input_holder == True:
+            data=request.data 
 
-        input=get_object_or_404(Input, pk=input_id)
+            input=get_object_or_404(Input, pk=input_id)
 
-        serializer=self.serializer_class(data=data, instance=input)
+            serializer=self.serializer_class(data=data, instance=input)
 
-        if serializer.is_valid():
-            serializer.save()
+            if serializer.is_valid():
+                serializer.save()
 
-            return Response(data=serializer.data, status=status.HTTP_200_OK)
+                return Response(data=serializer.data, status=status.HTTP_200_OK)
+        else:
+            return Response({"Error":"User must be an input user"})
         
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         
     @swagger_auto_schema(operation_summary="Delete an input")
     def delete(self, request, input_id):
-        input=get_object_or_404(Input, pk=input_id)
+        if request.user.is_input_holder == True:
+            input=get_object_or_404(Input, pk=input_id)
 
-        input.delete()
+            input.delete()
+        else:
+            return Response({"Error":"User must be an input user"})
 
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class UserInputView(generics.GenericAPIView):
     serializer_class=serializers.InputDetailViewSerializer
+    permission_classes=[IsAuthenticated,]
+
 
     @swagger_auto_schema(operation_summary="Get all inputs for a user")
     def get(self, request, user_id): 
-        user=User.objects.get(pk=user_id)
+        if request.user.is_input_holder == True:
+            user=User.objects.get(pk=user_id)
 
-        inputs=Input.objects.all().filter(input_holder=user)
+            inputs=Input.objects.all().filter(input_holder=user)
 
-        serializer=self.serializer_class(instance=inputs, many=True)
+            serializer=self.serializer_class(instance=inputs, many=True)
+        else:
+            return Response({"Error":"User must be an input user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
 
 class UserInputDetail(generics.GenericAPIView):
     serializer_class=serializers.InputDetailViewSerializer
+    permission_classes=[IsAuthenticated]
+
 
     @swagger_auto_schema(operation_summary="Get a user's specific input")
     def get(self, request, user_id, input_id):
-        user=User.objects.get(pk=user_id)
+        if request.user.is_input_holder == True:
+            user=User.objects.get(pk=user_id)
 
-        inputs=Input.objects.all().filter(input_holder=user).get(pk=input_id)
+            inputs=Input.objects.all().filter(input_holder=user).get(pk=input_id)
 
-        serializer=self.serializer_class(instance=inputs)
+            serializer=self.serializer_class(instance=inputs)
+        else:
+            return Response({"Error":"User must be an input user"})
 
+        
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
 
@@ -212,26 +269,31 @@ class InvestmentCreateListView(generics.GenericAPIView):
 
     @swagger_auto_schema(operation_summary="List all investments")
     def get(self, request):
+        if request.user.is_investor == True:
+            investments = Investment.objects.all()
 
-        investments = Investment.objects.all()
-
-        serializer=self.serializer_class(instance=investments, many=True)
+            serializer=self.serializer_class(instance=investments, many=True)
+        else:
+            return Response({"Error":"User must be an investor user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     @swagger_auto_schema(operation_summary="Create a new investment")
     def post (self, request):
-        data=request.data
+        if request.user.is_investor == True:
+            data=request.data
 
-        serializer=self.serializer_class(data=data)
+            serializer=self.serializer_class(data=data)
 
-        user=request.user
+            user=request.user
 
-        if serializer.is_valid():
-            serializer.save(investor_holder=user)
+            if serializer.is_valid():
+                serializer.save(investor_holder=user)
 
-            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
-        
+                return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response({"Error":"User must be an investor user"})
+
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -241,61 +303,80 @@ class InvestmentDetailView(generics.GenericAPIView):
 
     @swagger_auto_schema(operation_summary="Retrieve an investment by id")
     def get(self, request, investor_id):
-        
-        investment=get_object_or_404(Investment, pk=investor_id)
+        if request.user.is_investor == True:
 
-        serializer=self.serializer_class(instance=investment)
+            investment=get_object_or_404(Input, pk=investor_id)
 
+            serializer=self.serializer_class(instance=investment)
+        else:
+            return Response({"Error":"User must be an investor user"})
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     @swagger_auto_schema(operation_summary="Update an investmnent by id")
     def put(self, request, investor_id):
-        data=request.data 
+        if request.user.is_investor == True:
+            data=request.data 
 
-        investment=get_object_or_404(Investment, pk=investor_id)
+            investment=get_object_or_404(Investment, pk=investor_id)
 
-        serializer=self.serializer_class(data=data, instance=investment)
+            serializer=self.serializer_class(data=data, instance=investment)
 
-        if serializer.is_valid():
-            serializer.save()
+            if serializer.is_valid():
+                serializer.save()
 
-            return Response(data=serializer.data, status=status.HTTP_200_OK)
+                return Response(data=serializer.data, status=status.HTTP_200_OK)
+        else:
+            return Response({"Error":"User must be an investor user"})
         
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         
     @swagger_auto_schema(operation_summary="Delete an investment")
     def delete(self, request, investor_id):
-        investment=get_object_or_404(Investment, pk=investor_id)
+        if request.user.is_investor == True:
 
-        investment.delete()
+            investment=get_object_or_404(Investment, pk=investor_id)
+
+            investment.delete()
+        else:
+            return Response({"Error":"User must be an investor user"})
 
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class UserInvestmentView(generics.GenericAPIView):
     serializer_class=serializers.InvestDetailViewSerializer
+    permission_classes=[IsAuthenticated,]
+
 
     @swagger_auto_schema(operation_summary="Get all investments for a user")
-    def get(self, request, user_id): 
-        user=User.objects.get(pk=user_id)
+    def get(self, request, user_id):
+        if request.user.is_investor == True: 
+            user=User.objects.get(pk=user_id)
 
-        investments=Investment.objects.all().filter(investor_holder=user)
+            investments=Investment.objects.all().filter(investor_holder=user)
 
-        serializer=self.serializer_class(instance=investments, many=True)
+            serializer=self.serializer_class(instance=investments, many=True)
+        else:
+            return Response({"Error":"User must be an investor user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
 
 class UserInvestmentDetail(generics.GenericAPIView):
     serializer_class=serializers.InvestDetailViewSerializer
+    permission_classes=[IsAuthenticated,]
+
 
     @swagger_auto_schema(operation_summary="Get a user's specific investment")
     def get(self, request, user_id, investor_id):
-        user=User.objects.get(pk=user_id)
+        if request.user.is_investor == True: 
+            user=User.objects.get(pk=user_id)
 
-        investments=Investment.objects.all().filter(investor_holder=user).get(pk=investor_id)
+            investments=Investment.objects.all().filter(investor_holder=user).get(pk=investor_id)
 
-        serializer=self.serializer_class(instance=investments)
+            serializer=self.serializer_class(instance=investments)
+        else:
+            return Response({"Error":"User must be an investor user"})
 
         return Response(data=serializer.data, status=status.HTTP_200_OK)
